Move mesh_utils -> mesh_container
